refactor(extract): interpret case party semantics
